I worked on issues with pysha3. I think I have got it working in python 2.
I believe that sha3 is aliased to _sha3 in python3, so it may only work as is in python 2.

I changed all of the print statements to python 2 formatting.
I did this because I thought that is might be causing problems while running under python 2 in c9.
Before committing I commented out all of the print statements to avoid causing problems in python 3.